split txncontrol.
